[gpu] Upgrade sync point order numbers to 64-bit
The GPU SyncPointManager uses a 32-bit global sequence generator to
assign order numbers to scheduled tasks. This could theoretically wrap
around and cause subsequent sync-token waits to be bypassed due to
unsigned comparison logic.

This CL upgrades all variables, parameters, and fields representing sync
point order numbers to uint64_t. This ensures that a wraparound cannot
practically happen.

Also fixed a comment regarding scheduler task trace flow IDs.

Fixed: 516974902
Test: gpu_unittests --gtest_filter="*SyncPointManager*:*TaskGraph*:*Scheduler*"
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7910421
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1643970}

+++ gpu/command_buffer/service/task_graph.h
@@ -183,19 +183,19 @@ class GPU_COMMAND_BUFFER_SERVICE_EXPORT TaskGraph {
     }
 
     // Enqueues a task in the sequence and returns the generated order number.
-    uint32_t AddTask(TaskCallback task_callback,
+    uint64_t AddTask(TaskCallback task_callback,
                      std::vector<SyncToken> wait_fences,
                      const SyncToken& release,
                      ReportingCallback report_callback)
         EXCLUSIVE_LOCKS_REQUIRED(&TaskGraph::lock_);
 
-    virtual uint32_t AddTask(base::OnceClosure task_closure,
+    virtual uint64_t AddTask(base::OnceClosure task_closure,
                              std::vector<SyncToken> wait_fences,
                              const SyncToken& release,
                              ReportingCallback report_callback)
         EXCLUSIVE_LOCKS_REQUIRED(&TaskGraph::lock_);
 
-    virtual uint32_t BeginTask(base::OnceClosure* task_closure)
+    virtual uint64_t BeginTask(base::OnceClosure* task_closure)
         EXCLUSIVE_LOCKS_REQUIRED(&TaskGraph::lock_);
 
     // Should be called after running the closure returned by BeginTask().
@@ -214,7 +214,7 @@ class GPU_COMMAND_BUFFER_SERVICE_EXPORT TaskGraph {
     // - To avoid reentrancy, it is not called by AddTask() or FinishTask(),
     //   even if those methods result a new front task which is not blocked.
     // - It is called while holding `TaskGraph::lock_`.
-    virtual void OnFrontTaskUnblocked(uint32_t order_num)
+    virtual void OnFrontTaskUnblocked(uint64_t order_num)
         EXCLUSIVE_LOCKS_REQUIRED(&TaskGraph::lock_) {}
 
     // Sets the first dependency added time on the last task if it wasn't
@@ -228,7 +228,7 @@ class GPU_COMMAND_BUFFER_SERVICE_EXPORT TaskGraph {
 
     // Removes a waiting sync token fence.
     void RemoveWaitFence(const SyncToken& sync_token,
-                         uint32_t order_num,
+                         uint64_t order_num,
                          SequenceId release_sequence_id)
         EXCLUSIVE_LOCKS_REQUIRED(&TaskGraph::lock_);
 
@@ -255,7 +255,7 @@ class GPU_COMMAND_BUFFER_SERVICE_EXPORT TaskGraph {
       // `order_num`. The `sync_token` is released on the sequence identified
       // by `release_sequence_id`.
       WaitFence(const SyncToken& sync_token,
-                uint32_t order_num,
+                uint64_t order_num,
                 SequenceId release_sequence_id);
       WaitFence(WaitFence&& other);
 
@@ -264,7 +264,7 @@ class GPU_COMMAND_BUFFER_SERVICE_EXPORT TaskGraph {
       WaitFence& operator=(WaitFence&& other);
 
       SyncToken sync_token;
-      uint32_t order_num;
+      uint64_t order_num;
       SequenceId release_sequence_id;
 
       bool operator==(const WaitFence& other) const {
@@ -284,7 +284,7 @@ class GPU_COMMAND_BUFFER_SERVICE_EXPORT TaskGraph {
     struct Task {
       Task(Task&& other);
       Task(base::OnceClosure task_closure,
-           uint32_t order_num,
+           uint64_t order_num,
            const SyncToken& release,
            ReportingCallback report_callback);
       ~Task();
@@ -293,7 +293,7 @@ class GPU_COMMAND_BUFFER_SERVICE_EXPORT TaskGraph {
       // Always store tasks as closures. TaskCallbacks are bound with argument
       // and wrap as closures.
       base::OnceClosure task_closure;
-      uint32_t order_num;
+      uint64_t order_num;
       SyncToken release;
 
       ReportingCallback report_callback;
@@ -376,7 +376,7 @@ class GPU_COMMAND_BUFFER_SERVICE_EXPORT TaskGraph {
   };
 
   void SyncTokenFenceReleased(const SyncToken& sync_token,
-                              uint32_t order_num,
+                              uint64_t order_num,
                               SequenceId release_sequence_id,
                               SequenceId waiting_sequence_id)
       LOCKS_EXCLUDED(lock_);
